Added option for setting padding between modules.

+++ export/image/shapes/shapes.go
@@ -10,9 +10,10 @@ type DrawContext struct {
 	*gg.Context
 
 	X, Y          float64
-	Width, Height int
+	Width, Height float64
 
 	Color color.Color
 }
 
+// TODO: Access to matrix, gap info
 type ModuleShapeDrawer = func(ctx *DrawContext)
